+++ lib/message/batch/docs.go
@@ -10,6 +10,7 @@ func FieldSpec() docs.FieldSpec {
 Allows you to configure a [batching policy](/docs/configuration/batching).`,
 		Examples: []interface{}{
 			map[string]interface{}{
+				"count":     0,
 				"byte_size": 5000,
 				"period":    "1s",
 			},
@@ -18,19 +19,34 @@ Allows you to configure a [batching policy](/docs/configuration/batching).`,
 				"period": "1s",
 			},
 			map[string]interface{}{
+				"count":  0,
 				"period": "1m",
-				"condition": map[string]interface{}{
-					"bloblang": `this.contains("END BATCH")`,
-				},
+				"check":  `this.contains("END BATCH")`,
 			},
 		},
 		Children: docs.FieldSpecs{
-			docs.FieldCommon("count", "A number of messages at which the batch should be flushed. If `0` disables count based batching."),
-			docs.FieldCommon("byte_size", "An amount of bytes at which the batch should be flushed. If `0` disables size based batching."),
-			docs.FieldCommon("period", "A period in which an incomplete batch should be flushed regardless of its size.", "1s", "1m", "500ms"),
-			docs.FieldAdvanced("condition", "A [condition](/docs/components/conditions/about) to test against each message entering the batch, if this condition resolves to `true` then the batch is flushed."),
+			docs.FieldCommon(
+				"count",
+				"A number of messages at which the batch should be flushed. If `0` disables count based batching.",
+			),
+			docs.FieldCommon(
+				"byte_size",
+				"An amount of bytes at which the batch should be flushed. If `0` disables size based batching.",
+			),
+			docs.FieldCommon(
+				"period",
+				"A period in which an incomplete batch should be flushed regardless of its size.",
+				"1s", "1m", "500ms",
+			),
+			docs.FieldCommon(
+				"check",
+				"A [Bloblang query](/docs/guides/bloblang/about/) that should return a boolean value indicating whether a message should end a batch.",
+				`this.type == "end_of_transaction"`,
+			).HasDefault(""),
+			docs.FieldDeprecated("condition"),
 			docs.FieldAdvanced(
-				"processors", "A list of [processors](/docs/components/processors/about) to apply to a batch as it is flushed. This allows you to aggregate and archive the batch however you see fit. Please note that all resulting messages are flushed as a single batch, therefore splitting the batch into smaller batches using these processors is a no-op.",
+				"processors",
+				"A list of [processors](/docs/components/processors/about) to apply to a batch as it is flushed. This allows you to aggregate and archive the batch however you see fit. Please note that all resulting messages are flushed as a single batch, therefore splitting the batch into smaller batches using these processors is a no-op.",
 				[]map[string]interface{}{
 					{
 						"archive": map[string]interface{}{