+++ lib/processor/group_by.go
@@ -1,10 +1,13 @@
 package processor
 
 import (
+	"errors"
 	"fmt"
 	"strconv"
 	"time"
 
+	"github.com/Jeffail/benthos/v3/internal/bloblang"
+	"github.com/Jeffail/benthos/v3/internal/bloblang/mapping"
 	"github.com/Jeffail/benthos/v3/internal/docs"
 	"github.com/Jeffail/benthos/v3/lib/condition"
 	"github.com/Jeffail/benthos/v3/lib/log"
@@ -13,6 +16,7 @@ import (
 	"github.com/Jeffail/benthos/v3/lib/metrics"
 	"github.com/Jeffail/benthos/v3/lib/response"
 	"github.com/Jeffail/benthos/v3/lib/types"
+	"github.com/google/go-cmp/cmp"
 	olog "github.com/opentracing/opentracing-go/log"
 )
 
@@ -25,9 +29,9 @@ func init() {
 			CategoryComposition,
 		},
 		Summary: `
-Splits a batch of messages into N batches, where each resulting batch contains a group of messages determined by conditions that are applied per message of the original batch.`,
+Splits a [batch of messages](/docs/configuration/batching/) into N batches, where each resulting batch contains a group of messages determined by a [Bloblang query](/docs/guides/bloblang/about/).`,
 		Description: `
-Once the groups are established a list of processors are applied to their respective grouped batch, which can be used to label the batch as per their grouping.`,
+Once the groups are established a list of processors are applied to their respective grouped batch, which can be used to label the batch as per their grouping. Messages that do not pass the check of any specified group are placed in their own group.`,
 		Examples: []docs.AnnotatedExample{
 			{
 				Title:   "Grouped Processing",
@@ -36,8 +40,7 @@ Once the groups are established a list of processors are applied to their respec
 pipeline:
   processors:
     - group_by:
-      - condition:
-          bloblang: 'content().contains("this is a foo")'
+      - check: content().contains("this is a foo")
         processors:
           - archive:
               format: tar
@@ -48,7 +51,7 @@ pipeline:
 output:
   switch:
     cases:
-      - check: 'meta("grouping") == "foo"'
+      - check: meta("grouping") == "foo"
         output:
           gcp_pubsub:
             project: foo_prod
@@ -62,9 +65,13 @@ output:
 		},
 		FieldSpecs: docs.FieldSpecs{
 			docs.FieldCommon(
-				"condition",
-				"A [condition](/docs/components/conditions/about/) that is checked against each message individually in order to determine whether it belongs to a group.",
-			).HasDefault(map[string]interface{}{}),
+				"check",
+				"A [Bloblang query](/docs/guides/bloblang/about/) that should return a boolean value indicating whether a message belongs to a given group.",
+				`this.type == "foo"`,
+				`this.contents.urls.contains("https://benthos.dev/")`,
+				`true`,
+			).HasDefault(""),
+			docs.FieldDeprecated("condition"),
 			docs.FieldCommon(
 				"processors",
 				"A list of [processors](/docs/components/processors/about/) to execute on the newly formed group.",
@@ -73,22 +80,26 @@ output:
 		sanitiseConfigFunc: func(conf Config) (interface{}, error) {
 			groups := []interface{}{}
 			for _, g := range conf.GroupBy {
-				condSanit, err := condition.SanitiseConfig(g.Condition)
-				if err != nil {
-					return nil, err
-				}
 				procsSanit := []interface{}{}
 				for _, p := range g.Processors {
-					var procSanit interface{}
-					if procSanit, err = SanitiseConfig(p); err != nil {
+					procSanit, err := SanitiseConfig(p)
+					if err != nil {
 						return nil, err
 					}
 					procsSanit = append(procsSanit, procSanit)
 				}
-				groups = append(groups, map[string]interface{}{
-					"condition":  condSanit,
+				groupSanit := map[string]interface{}{
 					"processors": procsSanit,
-				})
+					"check":      g.Check,
+				}
+				if !isDefaultGroupCond(g.Condition) {
+					condSanit, err := condition.SanitiseConfig(g.Condition)
+					if err != nil {
+						return nil, err
+					}
+					groupSanit["condition"] = condSanit
+				}
+				groups = append(groups, groupSanit)
 			}
 			return groups, nil
 		},
@@ -98,10 +109,18 @@ output:
 
 //------------------------------------------------------------------------------
 
+func isDefaultGroupCond(cond condition.Config) bool {
+	if cond.Type == "" {
+		return true
+	}
+	return cmp.Equal(cond, condition.NewConfig())
+}
+
 // GroupByElement represents a group determined by a condition and a list of
 // group specific processors.
 type GroupByElement struct {
 	Condition  condition.Config `json:"condition" yaml:"condition"`
+	Check      string           `json:"check" yaml:"check"`
 	Processors []Config         `json:"processors" yaml:"processors"`
 }
 
@@ -121,6 +140,7 @@ func NewGroupByConfig() GroupByConfig {
 
 type group struct {
 	Condition  condition.Type
+	Check      *mapping.Executor
 	Processors []types.Processor
 }
 
@@ -151,12 +171,29 @@ func NewGroupBy(
 		nsLog := log.NewModule("." + groupPrefix)
 		nsStats := metrics.Namespaced(stats, groupPrefix)
 
-		if groups[i].Condition, err = condition.New(
-			gConf.Condition, mgr,
-			nsLog.NewModule(".condition"), metrics.Namespaced(nsStats, "condition"),
-		); err != nil {
-			return nil, fmt.Errorf("failed to create condition for group '%v': %v", i, err)
+		if !isDefaultGroupCond(gConf.Condition) {
+			if groups[i].Condition, err = condition.New(
+				gConf.Condition, mgr,
+				nsLog.NewModule(".condition"), metrics.Namespaced(nsStats, "condition"),
+			); err != nil {
+				return nil, fmt.Errorf("failed to create condition for group '%v': %v", i, err)
+			}
 		}
+
+		if len(gConf.Check) > 0 {
+			if groups[i].Check, err = bloblang.NewMapping("", gConf.Check); err != nil {
+				return nil, fmt.Errorf("failed to parse check for group '%v': %v", i, err)
+			}
+		}
+
+		if groups[i].Check == nil && groups[i].Condition == nil {
+			return nil, errors.New("a group definition must have a check query")
+		}
+
+		if groups[i].Check != nil && groups[i].Condition != nil {
+			return nil, errors.New("cannot specify both a condition and a check in a group")
+		}
+
 		for j, pConf := range gConf.Processors {
 			prefix := fmt.Sprintf("processor.%v", j)
 			var proc Type
@@ -207,16 +244,35 @@ func (g *GroupBy) ProcessMessage(msg types.Message) ([]types.Message, types.Resp
 
 	msg.Iter(func(i int, p types.Part) error {
 		for j, group := range g.groups {
-			if group.Condition.Check(message.Lock(msg, i)) {
-				groupStr := strconv.Itoa(j)
-				spans[i].LogFields(
-					olog.String("event", "grouped"),
-					olog.String("type", groupStr),
-				)
-				spans[i].SetTag("group", groupStr)
-				groups[j].Append(p.Copy())
-				g.mGroupPass[j].Incr(1)
-				return nil
+			if group.Condition != nil {
+				if group.Condition.Check(message.Lock(msg, i)) {
+					groupStr := strconv.Itoa(j)
+					spans[i].LogFields(
+						olog.String("event", "grouped"),
+						olog.String("type", groupStr),
+					)
+					spans[i].SetTag("group", groupStr)
+					groups[j].Append(p.Copy())
+					g.mGroupPass[j].Incr(1)
+					return nil
+				}
+			} else if group.Check != nil {
+				res, err := group.Check.QueryPart(i, msg)
+				if err != nil {
+					res = false
+					g.log.Errorf("Failed to test group %v: %v\n", j, err)
+				}
+				if res {
+					groupStr := strconv.Itoa(j)
+					spans[i].LogFields(
+						olog.String("event", "grouped"),
+						olog.String("type", groupStr),
+					)
+					spans[i].SetTag("group", groupStr)
+					groups[j].Append(p.Copy())
+					g.mGroupPass[j].Incr(1)
+					return nil
+				}
 			}
 		}